RTC#55592 64-bit support for iOS. Fix plugin warnings. Partial cherry-picks from commits 0d7bf3980516b577deae0ff7f365b5c53116cbc1 and 910f80fae8fe79eb8342bcd737b663c2d5f62d78 from cordova-plugin-contacts, commit 175f02ed6f19b4912eb345448d38d54c6c61d6e9 from cordova-plugin-geolocation, commits 007630da4866d867321008c6b8e1ebf57f05f259 and db9fa24cf8ba165aabae0460a9cbeb1dd1d67e62 from cordova-plugin-media-capture, commit 29fa80ba5dfa27f5be4e254442d72037cf3ff71e from cordova-plugin-media, commit 0f84287df03e7372eebe1d8620f8a78bb4900b4e from cordova-plugin-file-transfer, commits c9bab1f94c89e53c0ad8dbc77e1027eaa89527f1 and 91d6e10b29e492d8d72a2d5c759ebf7b876ecc3b from cordova-plugin-camera, commit 9c7d4442f4f74cba573cff8a65deeb4437f21840 from cordova-plugin-dialogs, commit d1b5e5bdebc954e149feecd745bdf9e4d5e2a668 from cordova-plugin-globalization

+++ CordovaLib/Classes/CDVCapture.m
@@ -420,7 +420,7 @@
         // NSLog(@"getFormatData: %@", [formatData description]);
     }
     if (bError) {
-        result = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageToErrorObject:errorCode];
+        result = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageToErrorObject:(int)errorCode];
     }
     if (result) {
         [self.commandDelegate sendPluginResult:result callbackId:callbackId];
@@ -787,7 +787,7 @@
 
     if (!self.pluginResult) {
         // return error
-        self.pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageToErrorObject:self.errorCode];
+        self.pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageToErrorObject:(int)self.errorCode];
     }
 
     self.avRecorder = nil;
